+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 - Warning: ksort() expects parameter 1 to be array, null given in FittingHelper
 - Search result is now sorted alphabetically
+- Fitting visualization is now responsive
 
 ## [v0.1-r20170719](https://github.com/ppfeufer/eve-online-fitting-manager/releases/tag/v0.1-r20170719) - 2017-07-19
 [Full Changelog](https://github.com/ppfeufer/eve-online-fitting-manager/compare/v0.1-r20170711...v0.1-r20170719)